moved tabs under ui
